+++ Paradigm/src/utopia/paradigm/shape/shape2d/Rectangle.scala
@@ -5,7 +5,7 @@ import utopia.flow.datastructure.template
 import utopia.flow.datastructure.template.Property
 import utopia.flow.generic.{DoubleType, FromModelFactory, ModelConvertible}
 import utopia.flow.generic.ValueConversions._
-import utopia.paradigm.generic.{PointType, Vector3DType}
+import utopia.paradigm.generic.{PointType, Vector2DType}
 import utopia.paradigm.generic.ParadigmValue._
 
 import scala.util.Try
@@ -20,26 +20,27 @@ object Rectangle extends FromModelFactory[Rectangle]
 	/**
 	  * A schema used when converting models to rectangles
 	  */
-	val schema = ModelDeclaration("topLeft" -> PointType, "top" -> Vector3DType, "leftEdgeLength" -> DoubleType)
+	val schema = ModelDeclaration("topLeft" -> PointType, "top" -> Vector2DType, "rightEdgeLength" -> DoubleType)
 	
-	override def apply(model: template.Model[Property]): Try[Rectangle] =
-	{
-		schema.validate(model).toTry.map { valid => Rectangle(valid("topLeft").getPoint, valid("top").getVector3D.in2D,
-			valid("leftEdgeLength").getDouble) }
+	override def apply(model: template.Model[Property]): Try[Rectangle] = {
+		schema.validate(model).toTry.map { valid =>
+			Rectangle(valid("topLeft").getPoint, valid("top").getVector2D, valid("rightEdgeLength").getDouble)
+		}
 	}
 }
 
 /**
   * Rectangles are 2D shapes that have four sides and where each corner is 90 degrees
-  * @param topLeft The top left corner of this rectangle
-  * @param top The top vector of this rectangle
-  * @param leftLength The length of the left edge of this rectangle
+  * @param topLeftCorner The top left corner of this rectangle
+  * @param topEdge The top vector of this rectangle
+  * @param rightEdgeLength The length of the right edge of this rectangle
   */
-case class Rectangle(topLeft: Point, top: Vector2D, leftLength: Double) extends Rectangular with ModelConvertible
+case class Rectangle(topLeftCorner: Point, topEdge: Vector2D, rightEdgeLength: Double)
+	extends Rectangular with ModelConvertible
 {
 	// ATTRIBUTES	-----------------
 	
-	override lazy val left = super.left
+	override lazy val rightEdge = super.rightEdge
 	
 	
 	// COMPUTED	---------------------
@@ -47,16 +48,15 @@ case class Rectangle(topLeft: Point, top: Vector2D, leftLength: Double) extends 
 	/**
 	  * @return The size of this rectangle (rotation lost)
 	  */
-	def toSize = Size(width, height)
-	
+	def toSize = Size(topEdge.length, rightEdgeLength)
 	/**
 	  * @return The bounds of this rectangle (rotation lost)
 	  */
-	def toBounds = Bounds(topLeft, toSize)
+	def toBounds = Bounds(topLeftCorner, toSize)
 	
 	
 	// IMPLEMENTED	--------------------
 	
 	override def toModel =
-		Model(Vector(("topLeft", topLeft), ("top", top.in3D), ("leftEdgeLength", leftLength)))
+		Model(Vector(("topLeft", topLeftCorner), ("top", topEdge), ("rightEdgeLength", rightEdgeLength)))
 }